requirements.txt
Django==4.2
Pillow==9.5.0
django-admin-sortable2==2.1.3
django-tinymce==3.5.0